Renaming _version module to __version__
